% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_util.R
\name{generate_COVID_animation}
\alias{generate_COVID_animation}
\title{Generate_COVID_animation}
\usage{
generate_COVID_animation(
  Input_data = Cases_final,
  Total_ID = "USA_total",
  states_selected = c("New York", "California", "Washington")
)
}
\arguments{
\item{Total_ID}{This is the entry which will track the total cases}

\item{states_selected}{A vector of special states to track cases in}

\item{Imput_data}{Dataset with Day, cumulative cases, states}
}
\value{
An object that can be used within gganimate to generate the gif.
}
\description{
Generate a gif of COVID 19 visualizations.
}
\author{
Kanishka Narayan
}
